@@ -5,17 +5,17 @@ description: A sample command-line application.
 # author: long1eu <lung.razvan@gmail.com>
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.9.0 <3.0.0'
 
 dependencies:
-  email_validator: ^1.0.4
+  email_validator: ^1.0.5
   intl: ^0.16.1
-  uuid: ^2.0.4
-  crypto: ^2.1.4
-  hive: ^1.4.1+1
+  uuid: ^2.2.2
+  crypto: ^2.1.5
+  hive: ^1.4.4
   firebase_auth_vm:
     path: ../../
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.12.0
+  pedantic: ^1.0.0
+  test: ^1.15.3
