@@ -6,8 +6,7 @@ part of firebase_auth_example;
 
 Future<EmailPasswordAuthCredential> _getEmailPasswordAuthCredential() async {
   final EmailAndPassword result = await getEmailAndPassword();
-  return EmailAuthProvider.getCredential(
-      email: result.email, password: result.password);
+  return EmailAuthProvider.getCredential(email: result.email, password: result.password);
 }
 
 Future<PhoneAuthCredential> _getPhoneAuthCredential() async {
@@ -29,17 +28,19 @@ Future<PhoneAuthCredential> _getPhoneAuthCredential() async {
   final String verificationId = await FirebaseAuth.instance.verifyPhoneNumber(
     phoneNumber: phoneNumber,
     presenter: (Uri uri) {
-      console
-        ..println(
-            'In order to verify the app please complete the recaptcha change by clicking the link below.')
-        ..println(uri.toString().bold.cyan.reset);
+      console..println(
+          'In order to verify the app please complete the recaptcha change by clicking the link below.')..println(uri
+          .toString()
+          .bold
+          .cyan
+          .reset);
     },
   );
 
   console.println();
   option = StringOption(
     question:
-        'Great! A SMS was sent to ${phoneNumber.yellow.reset}. Please type the ${'code'.yellow.reset} you receive.',
+    'Great! A SMS was sent to ${phoneNumber.yellow.reset}. Please type the ${'code'.yellow.reset} you receive.',
     fieldBuilder: () => 'code: ',
     validator: (String response) {
       if (response.isEmpty) {
@@ -51,8 +52,7 @@ Future<PhoneAuthCredential> _getPhoneAuthCredential() async {
   );
 
   final String code = await option.show();
-  return PhoneAuthProvider.getCredential(
-      verificationId: verificationId, verificationCode: code);
+  return PhoneAuthProvider.getCredential(verificationId: verificationId, verificationCode: code);
 }
 
 Future<GoogleAuthCredential> _getGoogleAuthCredential() async {
@@ -85,8 +85,7 @@ Future<GoogleAuthCredential> _getGoogleAuthCredential() async {
       );
     },
     codePollValidator: (Map<String, dynamic> pollResponse) {
-      if (pollResponse['access_token'] != null &&
-          pollResponse['id_token'] != null) {
+      if (pollResponse['access_token'] != null && pollResponse['id_token'] != null) {
         return null;
       } else if (pollResponse['error'] == 'access_denied') {
         return 'You canceled the authorization.';
@@ -102,8 +101,7 @@ Future<GoogleAuthCredential> _getGoogleAuthCredential() async {
   final String accessToken = credentials['access_token'];
   final String idToken = credentials['id_token'];
 
-  return GoogleAuthProvider.getCredential(
-      idToken: idToken, accessToken: accessToken);
+  return GoogleAuthProvider.getCredential(idToken: idToken, accessToken: accessToken);
 }
 
 Future<FacebookAuthCredential> _getFacebookAuthCredential() async {
@@ -139,7 +137,7 @@ Future<FacebookAuthCredential> _getFacebookAuthCredential() async {
       }
       final Map<String, dynamic> error = pollResponse['error'];
       switch (error['error_subcode']) {
-        // pending
+      // pending
         case 1349174:
           return '';
         default:
@@ -167,8 +165,7 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   );
 
   console //
-    ..println('Visit this link and login with GitHub')
-    ..println(uri);
+    ..println('Visit this link and login with GitHub')..println(uri);
 
   final HttpRequest request = await server.first;
   final Map<String, dynamic> data = request.requestedUri.queryParameters;
@@ -176,8 +173,7 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   final String receivedState = data['state'];
 
   if (request.method != 'GET') {
-    throw Exception(
-        'Invalid response from server (expected GET request callback, got: ${request.method}).');
+    throw Exception('Invalid response from server (expected GET request callback, got: ${request.method}).');
   }
   if (receivedState != state) {
     throw StateError('The requested state doesn\'t match.');
@@ -191,7 +187,8 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   await request.response.close();
   await server.close();
 
-  final Progress progress = Progress('Verifying credentials')..show();
+  final Progress progress = Progress('Verifying credentials')
+    ..show();
   final Response response = await post(
     'https://us-central1-flutter-sdk.cloudfunctions.net/handler',
     headers: <String, String>{'content-type': 'application/json'},
@@ -206,8 +203,7 @@ Future<GithubAuthCredential> _getGithubAuthCredential() async {
   );
   await progress.cancel();
 
-  final Map<String, dynamic> accessTokenResponse =
-      jsonDecode(response.body)['result'];
+  final Map<String, dynamic> accessTokenResponse = jsonDecode(response.body)['result'];
   if (response.statusCode > 200) {
     throw StateError(accessTokenResponse['error']);
   }
@@ -224,16 +220,15 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
     accessTokenSecret: _twitterAccessTokenSecret,
   );
 
-  Progress progress = Progress('Getting Twitter configuration')..show();
+  Progress progress = Progress('Getting Twitter configuration')
+    ..show();
   await client.initialize();
   await progress.cancel();
   final TwitterRequestToken requestToken = client.requestToken;
   final HttpServer server = client.server;
 
-  console
-    ..println('Visit this link and login with Twitter')
-    ..println(
-        'https://api.twitter.com/oauth/authenticate?oauth_token=${requestToken.token}');
+  console..println('Visit this link and login with Twitter')..println(
+      'https://api.twitter.com/oauth/authenticate?oauth_token=${requestToken.token}');
 
   final HttpRequest request = await server.first;
   final Map<String, dynamic> data = request.requestedUri.queryParameters;
@@ -241,8 +236,7 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
   final String oauthVerifier = data['oauth_verifier'];
 
   if (request.method != 'GET') {
-    throw Exception(
-        'Invalid response from server (expected GET request callback, got: ${request.method}).');
+    throw Exception('Invalid response from server (expected GET request callback, got: ${request.method}).');
   }
   if (requestToken.token != oauthToken) {
     throw StateError('The request doesn\'t match.');
@@ -256,9 +250,9 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
   await request.response.close();
   await server.close();
 
-  progress = Progress('Validating credentials')..show();
-  final Response response =
-      await client.post('oauth/access_token', headers: <String, String>{
+  progress = Progress('Validating credentials')
+    ..show();
+  final Response response = await client.post('oauth/access_token', headers: <String, String>{
     'oauth_token': requestToken.token,
     'oauth_verifier': oauthVerifier,
   });
@@ -272,8 +266,7 @@ Future<TwitterAuthCredential> _getTwitterAuthCredential() async {
   final String userAuthToken = queryParameters['oauth_token'];
   final String userAuthTokenSecret = queryParameters['oauth_token_secret'];
 
-  return TwitterAuthProvider.getCredential(
-      authToken: userAuthToken, authTokenSecret: userAuthTokenSecret);
+  return TwitterAuthProvider.getCredential(authToken: userAuthToken, authTokenSecret: userAuthTokenSecret);
 }
 
 Future<OAuthCredential> _getYahooAuthCredential() {
@@ -317,8 +310,7 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   );
 
   console //
-    ..println('Visit this link and login with $providerName')
-    ..println(authorizationUri);
+    ..println('Visit this link and login with $providerName')..println(authorizationUri);
 
   final HttpRequest request = await server.first;
   final Map<String, dynamic> data = request.requestedUri.queryParameters;
@@ -326,8 +318,7 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   final String receivedState = data['state'];
 
   if (request.method != 'GET') {
-    throw Exception(
-        'Invalid response from server (expected GET request callback, got: ${request.method}).');
+    throw Exception('Invalid response from server (expected GET request callback, got: ${request.method}).');
   }
   if (receivedState != state) {
     throw StateError('The requested state doesn\'t match.');
@@ -341,7 +332,8 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   await request.response.close();
   await server.close();
 
-  final Progress progress = Progress('Verifying credentials')..show();
+  final Progress progress = Progress('Verifying credentials')
+    ..show();
   final Response response = await post(
     'https://us-central1-flutter-sdk.cloudfunctions.net/handler',
     headers: <String, String>{'content-type': 'application/json'},
@@ -357,24 +349,21 @@ Future<OAuthCredential> _getOAuthAuthCredential({
   );
   await progress.cancel();
 
-  final Map<String, dynamic> accessTokenResponse =
-      jsonDecode(response.body)['result'] ?? jsonDecode(response.body);
+  final Map<String, dynamic> accessTokenResponse = jsonDecode(response.body)['result'] ?? jsonDecode(response.body);
   if (response.statusCode > 200 || accessTokenResponse.containsKey('error')) {
     print(response.body);
     throw StateError(accessTokenResponse['error'].toString());
   }
 
   final String accessToken = accessTokenResponse['access_token'];
-  return OAuthProvider.getCredentialWithAccessToken(
-      providerId: providerId, accessToken: accessToken);
+  return OAuthProvider.getCredentialWithAccessToken(providerId: providerId, accessToken: accessToken);
 }
 
-Future<AuthResult> _presentSignInWithCredential(
-    AuthCredential credential) async {
+Future<AuthResult> _presentSignInWithCredential(AuthCredential credential) async {
   console.println();
-  final Progress progress = Progress('Siging in')..show();
-  final AuthResult result =
-      await FirebaseAuth.instance.signInWithCredential(credential);
+  final Progress progress = Progress('Siging in')
+    ..show();
+  final AuthResult result = await FirebaseAuth.instance.signInWithCredential(credential);
   await progress.cancel();
   console.clearScreen();
   return result;
