@@ -4,115 +4,127 @@
 
 part of firebase_auth_vm;
 
-abstract class SignInWithGameCenterRequest
-    implements
-        Built<SignInWithGameCenterRequest, SignInWithGameCenterRequestBuilder> {
-  factory SignInWithGameCenterRequest({
-    @required String playerId,
-    @required String publicKeyUrl,
-    @required Uint8List signature,
-    @required Uint8List salt,
-    @required DateTime timestamp,
-    String accessToken,
-    @required String displayName,
-  }) {
-    final String encodedSignature = signature != null
-        ? base64Encode(signature)
-            .replaceAll('/', '_')
-            .replaceAll('+', '-')
-            .replaceAll('=', '')
-        : null;
-
-    final String encodedSalt = signature != null //
-        ? base64Encode(salt)
-            .replaceAll('/', '_')
-            .replaceAll('+', '-')
-            .replaceAll('=', '')
-        : null;
-
-    return _$SignInWithGameCenterRequest(
-        (SignInWithGameCenterRequestBuilder b) {
-      b
-        ..playerId = playerId
-        ..publicKeyUrl = publicKeyUrl
-        ..signature = encodedSignature
-        ..salt = encodedSalt
-        ..timestamp = timestamp?.toUtc()?.millisecondsSinceEpoch
-        ..accessToken = accessToken
-        ..displayName = displayName;
-    });
-  }
-
-  SignInWithGameCenterRequest._();
+/// The request to sign in with Game Center account
+class SignInWithGameCenterRequest {
+  SignInWithGameCenterRequest({
+    @required this.playerId,
+    @required this.publicKeyUrl,
+    @required this.signature,
+    @required this.salt,
+    @required this.timestamp,
+    this.accessToken,
+    @required this.displayName,
+  });
 
   /// The playerID to verify.
-  @nullable
-  String get playerId;
+  /*@nullable*/
+  final String playerId;
 
   /// The URL for the public encryption key.
-  @nullable
-  String get publicKeyUrl;
+  /*@nullable*/
+  final String publicKeyUrl;
 
   /// The verification signature data generated by Game Center.
-  @nullable
-  String get signature;
+  /*@nullable*/
+  final Uint8List signature;
 
   /// A random strong used to compute the hash and keep it randomized.
-  @nullable
-  String get salt;
+  /*@nullable*/
+  final Uint8List salt;
 
   /// The date and time that the signature was created.
-  @nullable
-  int get timestamp;
+  /*@nullable*/
+  final DateTime timestamp;
 
   /// The STS Access Token for the authenticated user, only needed for linking the user.
-  @nullable
-  @BuiltValueField(wireName: 'idToken')
-  String get accessToken;
+  /*@nullable*/
+  final String accessToken;
 
   /// The display name of the local Game Center player.
-  @nullable
-  String get displayName;
+  /*@nullable*/
+  final String displayName;
 
-  Map<String, dynamic> get json => serializers.serializeWith(serializer, this);
+  Map<String, dynamic> get json {
+    return <String, dynamic>{
+      'playerId': playerId,
+      'publicKeyUrl': publicKeyUrl,
+      'signature': base64UrlEncode(signature),
+      'salt': base64UrlEncode(salt),
+      'timestamp': timestamp?.toUtc()?.millisecondsSinceEpoch,
+      'idToken': accessToken,
+      'displayName': displayName,
+    };
+  }
 
-  static Serializer<SignInWithGameCenterRequest> get serializer =>
-      _$signInWithGameCenterRequestSerializer;
+  @override
+  String toString() {
+    return (ToStringHelper(SecureTokenRequest)
+          ..add('playerId', playerId)
+          ..add('publicKeyUrl', publicKeyUrl)
+          ..add('signature', base64UrlEncode(signature))
+          ..add('salt', base64UrlEncode(salt))
+          ..add('timestamp', timestamp)
+          ..add('accessToken', accessToken)
+          ..add('displayName', displayName))
+        .toString();
+  }
 }
 
-abstract class SignInWithGameCenterResponse
-    implements
-        Built<SignInWithGameCenterResponse,
-            SignInWithGameCenterResponseBuilder> {
-  factory SignInWithGameCenterResponse() = _$SignInWithGameCenterResponse;
+class SignInWithGameCenterResponse {
+  factory SignInWithGameCenterResponse.fromJson(Map<dynamic, dynamic> json) {
+    return SignInWithGameCenterResponse._(
+      idToken: json['idToken'],
+      refreshToken: json['refreshToken'],
+      localId: json['localId'],
+      playerId: json['playerId'],
+      expiresIn: json['expiresIn'],
+      isNewUser: json['isNewUser'],
+      displayName: json['displayName'],
+    );
+  }
 
-  factory SignInWithGameCenterResponse.fromJson(Map<dynamic, dynamic> json) =>
-      serializers.deserializeWith(serializer, json);
-
-  SignInWithGameCenterResponse._();
+  SignInWithGameCenterResponse._({
+    this.idToken,
+    this.refreshToken,
+    this.localId,
+    this.playerId,
+    this.expiresIn,
+    this.isNewUser,
+    this.displayName,
+  });
 
   /// Either an authorization code suitable for performing an STS token exchange, or the access token from Secure Token
   /// Service, depending on whether [returnSecureToken] is set on the request.
-  String get idToken;
+  final String idToken;
 
   /// The refresh token from Secure Token Service.
-  String get refreshToken;
+  final String refreshToken;
 
   /// The Firebase Auth user ID.
-  String get localId;
+  final String localId;
 
   /// The verified player ID.
-  String get playerId;
+  final String playerId;
 
   /// The approximate expiration date of the access token.
-  String get expiresIn;
+  final String expiresIn;
 
   /// Flag indicating that the user signing in is a new user and not a returning user.
-  bool get isNewUser;
+  final bool isNewUser;
 
   /// The user's Game Center display name.
-  String get displayName;
+  final String displayName;
 
-  static Serializer<SignInWithGameCenterResponse> get serializer =>
-      _$signInWithGameCenterResponseSerializer;
+  @override
+  String toString() {
+    return (ToStringHelper(SecureTokenRequest)
+          ..add('idToken', idToken)
+          ..add('refreshToken', refreshToken)
+          ..add('localId', localId)
+          ..add('playerId', playerId)
+          ..add('expiresIn', expiresIn)
+          ..add('isNewUser', isNewUser)
+          ..add('displayName', displayName))
+        .toString();
+  }
 }
