@@ -29,23 +29,24 @@ class ApiKeyClient extends DelegatingClient {
     }
 
     if (url.queryParameters.isEmpty) {
-      url =
-          url.replace(queryParameters: <String, String>{'key': _encodedApiKey});
+      url = url.replace(queryParameters: <String, String>{'key': _encodedApiKey});
     } else {
-      url = url.replace(queryParameters: <String, String>{
-        ...url.queryParameters,
-        'key': _encodedApiKey
-      });
+      url = url.replace(queryParameters: <String, String>{...url.queryParameters, 'key': _encodedApiKey});
     }
 
-    final RequestImpl modifiedRequest =
-        RequestImpl(request.method, url, request.finalize());
-    final Map<String, String> headers = <String, String>{
+    final RequestImpl modifiedRequest = RequestImpl(request.method, url, request.finalize());
+
+    Map<String, String> headers;
+    if (_headers != null) {
+      headers = await _headers();
+      headers ??= <String, String>{};
+    }
+
+    headers = <String, String>{
       ...request.headers,
-      if (_headers != null) ..._headers() ?? <String, String>{},
+      ...headers,
       if (locale != null) 'X-Firebase-Locale': locale,
-      'X-Client-Version':
-          '${Platform.operatingSystem}/FirebaseSDK/$sdkVersion/dart',
+      'X-Client-Version': '${Platform.operatingSystem}/FirebaseSDK/$sdkVersion/dart',
     };
     modifiedRequest.headers.addAll(headers);
 
@@ -102,8 +103,7 @@ Future<StreamedResponse> _validateResponse(StreamedResponse response) async {
     // Some error happened, try to decode the response and fetch the error.
     final Stream<String> stringStream = _decodeStreamAsText(response);
     if (stringStream != null) {
-      final dynamic jsonResponse =
-          await stringStream.transform(json.decoder).first;
+      final dynamic jsonResponse = await stringStream.transform(json.decoder).first;
       if (jsonResponse is Map && jsonResponse['error'] is Map) {
         final Map<dynamic, dynamic> error = jsonResponse['error'];
         final dynamic codeValue = error['code'];
@@ -124,8 +124,7 @@ Stream<String> _decodeStreamAsText(StreamedResponse response) {
   // TODO(long1eu): Correctly handle the response content-types, using correct decoder.
   // Currently we assume that the api endpoint is responding with json encoded in UTF8.
   final String contentType = response.headers['content-type'];
-  if (contentType != null &&
-      contentType.toLowerCase().startsWith('application/json')) {
+  if (contentType != null && contentType.toLowerCase().startsWith('application/json')) {
     return response.stream.transform(const Utf8Decoder(allowMalformed: true));
   } else {
     return null;
